requirements.txt
keras
sklearn
numpy
matplotlib.pyplot
cartopy.crs